Create seperate file for Icon enum
