+++ Weather App/CurrentWeather.swift
@@ -9,52 +9,6 @@
 import Foundation
 import UIKit
 
-// Icon possibilities
-enum Icon: String {
-    
-    case ClearDay = "clear-day"
-    case ClearNight = "clear-night"
-    case Rain = "rain"
-    case Snow = "snow"
-    case Sleet = "sleet"
-    case Wind = "wind"
-    case Fog = "fog"
-    case Cloudy = "cloudy"
-    case PartlyCloudyDay = "partly-cloudy-day"
-    case PartlyCloudyNight = "partly-cloudy-night"
-    
-    func toImage() -> UIImage? {
-        var imageName: String
-        
-        // Enums can be initilized with passing in a raw value
-        // If the value matches one of the enum's values, we get it back
-        switch self {
-        case .ClearDay:
-            imageName = "clear-day.png"
-        case .ClearNight:
-            imageName = "clear-night.png"
-        case .Rain:
-            imageName = "rain.png"
-        case .Snow:
-            imageName = "snow.png"
-        case .Sleet:
-            imageName = "sleet.png"
-        case .Wind:
-            imageName = "wind.png"
-        case .Fog:
-            imageName = "fog.png"
-        case .Cloudy:
-            imageName = "cloudy.png"
-        case .PartlyCloudyDay:
-            imageName = "cloudy-day.png"
-        case .PartlyCloudyNight:
-            imageName = "cloudy-night.png"
-        }
-        
-        return UIImage(named: imageName)
-    }
-}
-
 struct CurrentWeather {
     
     let temperature: Int?
